refactor all appointment classes
